pyproject.toml
[project]
name = "adk-slack-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "google-adk>=1.2.1",
    "slack-bolt>=1.23.0",
]
